Introduce `ChatMessageAttachmentInfoView`
